refactor: move theme to separate file
